+++ CHANGELOG.md
@@ -83,6 +83,7 @@ data_source_fiat:
 - Bitfinex parser: new "movements" data file format added.
 - Accounting tool: new transaction types Lost and Airdrop added.
 - Binance parser: added trades statement export format.
+- Conversion tool: added parser for HecoInfo explorer.
 ### Changed
 - Conversion tool: UnknownAddressError exception changed to generic DataFilenameError.
 - Binance parser: use filename to determine if deposits or withdrawals.